To address is sent in a hidden field now
Putting the email address in the URL in the form action attribute was a cute idea, but too many email addresses weren't valid in a URL for it work well enough in my opinion. I'm particularly referring to the + sign addresses you can use to create ad-hock sibling addresses in Gmail, which I was using for testing.